[velbus] Made constants and handler packages internal
