+++ addons/binding/org.openhab.binding.velbus/src/main/java/org/openhab/binding/velbus/internal/VelbusHandlerFactory.java
@@ -8,7 +8,7 @@
  */
 package org.openhab.binding.velbus.internal;
 
-import static org.openhab.binding.velbus.VelbusBindingConstants.*;
+import static org.openhab.binding.velbus.internal.VelbusBindingConstants.*;
 
 import java.util.HashMap;
 import java.util.Hashtable;
@@ -22,15 +22,15 @@ import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
-import org.openhab.binding.velbus.handler.VelbusBlindsHandler;
-import org.openhab.binding.velbus.handler.VelbusBridgeHandler;
-import org.openhab.binding.velbus.handler.VelbusDimmerHandler;
-import org.openhab.binding.velbus.handler.VelbusRelayHandler;
-import org.openhab.binding.velbus.handler.VelbusSensorHandler;
-import org.openhab.binding.velbus.handler.VelbusVMBGPHandler;
-import org.openhab.binding.velbus.handler.VelbusVMBGPOHandler;
-import org.openhab.binding.velbus.handler.VelbusVMBPIROHandler;
 import org.openhab.binding.velbus.internal.discovery.VelbusThingDiscoveryService;
+import org.openhab.binding.velbus.internal.handler.VelbusBlindsHandler;
+import org.openhab.binding.velbus.internal.handler.VelbusBridgeHandler;
+import org.openhab.binding.velbus.internal.handler.VelbusDimmerHandler;
+import org.openhab.binding.velbus.internal.handler.VelbusRelayHandler;
+import org.openhab.binding.velbus.internal.handler.VelbusSensorHandler;
+import org.openhab.binding.velbus.internal.handler.VelbusVMBGPHandler;
+import org.openhab.binding.velbus.internal.handler.VelbusVMBGPOHandler;
+import org.openhab.binding.velbus.internal.handler.VelbusVMBPIROHandler;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Component;
