+++ addons/binding/org.openhab.binding.velbus/src/main/java/org/openhab/binding/velbus/internal/handler/VelbusVMBPIROHandler.java
@@ -6,9 +6,9 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.velbus.handler;
+package org.openhab.binding.velbus.internal.handler;
 
-import static org.openhab.binding.velbus.VelbusBindingConstants.*;
+import static org.openhab.binding.velbus.internal.VelbusBindingConstants.THING_TYPE_VMBPIRO;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -19,16 +19,15 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 
 /**
- * The {@link VelbusVMBGPHandler} is responsible for handling commands, which are
+ * The {@link VelbusVMBPIROHandler} is responsible for handling commands, which are
  * sent to one of the channels.
  *
  * @author Cedric Boon - Initial contribution
  */
-public class VelbusVMBGPHandler extends VelbusTemperatureSensorHandler {
-    public static final Set<ThingTypeUID> SUPPORTED_THING_TYPES = new HashSet<>(Arrays.asList(THING_TYPE_VMBGP1,
-            THING_TYPE_VMBGP2, THING_TYPE_VMBGP4, THING_TYPE_VMBGP4, THING_TYPE_VMBGP4PIR));
+public class VelbusVMBPIROHandler extends VelbusTemperatureSensorHandler {
+    public static final Set<ThingTypeUID> SUPPORTED_THING_TYPES = new HashSet<>(Arrays.asList(THING_TYPE_VMBPIRO));
 
-    public VelbusVMBGPHandler(Thing thing) {
-        super(thing, 4, new ChannelUID(thing.getUID(), "CH9"));
+    public VelbusVMBPIROHandler(Thing thing) {
+        super(thing, 0, new ChannelUID(thing.getUID(), "CH9"));
     }
 }
